+++ src/LDL/Validators/Chain/AndValidatorChain.php
@@ -76,7 +76,8 @@ class AndValidatorChain extends AbstractValidatorChain
 
     public static function fromConfig(
         ValidatorConfigInterface $config,
-        iterable $validators=null
+        iterable $validators=null,
+        string $description=null
     ): ValidatorChainInterface
     {
         if(!$config instanceof Config\ValidatorChainConfig){
@@ -92,7 +93,7 @@ class AndValidatorChain extends AbstractValidatorChain
             $validators,
             $config->isDumpable(),
             $config->isNegated(),
-            $config->getDescription()
+            $description
         );
     }
 }