+++ src/LDL/Validators/Chain/Dumper/ValidatorChainExprDumper.php
@@ -5,6 +5,7 @@ namespace LDL\Validators\Chain\Dumper;
 use LDL\Framework\Helper\IterableHelper;
 use LDL\Validators\Chain\ValidatorChainInterface;
 use LDL\Validators\Config\NegatedValidatorConfigInterface;
+use LDL\Validators\ValidatorHasConfigInterface;
 use LDL\Validators\ValidatorInterface;
 
 class ValidatorChainExprDumper implements ValidatorChainDumperInterface
@@ -27,6 +28,11 @@ class ValidatorChainExprDumper implements ValidatorChainDumperInterface
                 }
 
                 $class = get_class($validator);
+
+                if(!$validator instanceof ValidatorHasConfigInterface){
+                    return $class;
+                }
+
                 $config = $validator->getConfig();
 
                 if($config instanceof NegatedValidatorConfigInterface){