+++ src/LDL/Validators/InterfaceComplianceValidator.php
@@ -5,6 +5,7 @@ namespace LDL\Validators;
 use LDL\Validators\Config\InterfaceComplianceValidatorConfig;
 use LDL\Validators\Config\ValidatorConfigInterface;
 use LDL\Validators\Exception\TypeMismatchException;
+use LDL\Validators\Traits\ValidatorDescriptionTrait;
 use LDL\Validators\Traits\ValidatorHasConfigInterfaceTrait;
 use LDL\Validators\Traits\ValidatorValidateTrait;
 
@@ -12,10 +13,14 @@ class InterfaceComplianceValidator implements ValidatorInterface, NegatedValidat
 {
     use ValidatorValidateTrait {validate as _validate;}
     use ValidatorHasConfigInterfaceTrait;
+    use ValidatorDescriptionTrait;
+
+    private const DESCRIPTION = 'Validate interface';
 
     public function __construct(string $interface, bool $negated=false, bool $dumpable=true, string $description=null)
     {
-        $this->_tConfig = new InterfaceComplianceValidatorConfig($interface, $negated, $dumpable, $description);
+        $this->_tConfig = new InterfaceComplianceValidatorConfig($interface, $negated, $dumpable);
+        $this->_tDescription = $description ?? self::DESCRIPTION;
     }
 
     /**
@@ -70,7 +75,13 @@ class InterfaceComplianceValidator implements ValidatorInterface, NegatedValidat
         throw new TypeMismatchException($msg);
     }
 
-    public static function fromConfig(ValidatorConfigInterface $config): ValidatorInterface
+    /**
+     * @param ValidatorConfigInterface $config
+     * @param string|null $description
+     * @return ValidatorInterface
+     * @throws TypeMismatchException
+     */
+    public static function fromConfig(ValidatorConfigInterface $config, string $description=null): ValidatorInterface
     {
         if(false === $config instanceof InterfaceComplianceValidatorConfig){
             $msg = sprintf(
@@ -84,6 +95,6 @@ class InterfaceComplianceValidator implements ValidatorInterface, NegatedValidat
         /**
          * @var InterfaceComplianceValidatorConfig $config
          */
-        return new self($config->getInterface(), $config->isNegated(), $config->isDumpable());
+        return new self($config->getInterface(), $config->isNegated(), $config->isDumpable(), $description);
     }
 }